Moving ES-related code from core to sfeos_helpers.
