+++ stac_fastapi/sfeos_helpers/stac_fastapi/sfeos_helpers/models/patch.py
